+++ BussinessLayer/DTOs/ModuloInventario/Almacenes/InvInventarioSucursalReponse.cs
@@ -7,8 +7,8 @@ namespace BussinessLayer.DTOs.ModuloInventario.Almacenes
     {
         public int Id { get; set; }
         public int IdProducto { get; set; }
-        public int IdEmpresa { get; set; }
-        public int IdSucursal { get; set; }
+        public long IdEmpresa { get; set; }
+        public long IdSucursal { get; set; }
         public string UbicacionExhibicion { get; set; }
         public string UbicacionGuardada { get; set; }
         public int CantidadProducto { get; set; }
