+++ BussinessLayer/DTOs/ModuloInventario/Almacenes/InvMovInventarioSucursalReponse.cs
@@ -6,7 +6,7 @@ namespace BussinessLayer.DTOs.ModuloInventario.Almacenes
     public class InvMovInventarioSucursalReponse: AuditableEntitiesReponse
     {
         public int Id { get; set; }
-        public int IdSucursal { get; set; }
+        public long IdSucursal { get; set; }
         public int IdTransaccion { get; set; }
         public int IdModulo { get; set; }
         public int IdTransaccionOrigen { get; set; }
